chore: Move util functions from root.go into utils.go
